Player.cs 에 인벤토리 프로퍼티, 레벨업, AddGold 추가

+++ NotepadKnights/Player/PlayerUI.cs
@@ -69,7 +69,7 @@ namespace NotepadKnights
                 playerInput.ScreenChanges();
         }
         // 공격 이후 UI
-        public void DisplayAttackResult(int playerDamage, string msg)
+        public void DisplayAttackResult(float playerDamage, string msg)
         {
             Console.Clear();
